Async fhir request handler

+++ docker/service/src/main/java/uk/nhs/adaptors/scr/clients/SpineClient.java
@@ -92,7 +92,7 @@ public class SpineClient {
         var request = new HttpGet(url);
         //TODO: set headers
 
-        int hardTimeout = spineConfiguration.getScrResultHardTimeout();
+        long hardTimeout = spineConfiguration.getScrResultHardTimeout();
         TimerTask abortRequestTask = new TimerTask() {
             @Override
             public void run() {
